Segment out abi from main votium index.js
Segment out abi from main votium index.js
TVL = CVX + cvxCRV locked in contract
